Correct diagnostic ID prefix, fix race condition due to fixtures and project reference to Altinn.App.Core, some moving around
