+++ src/Altinn.App.Analyzers/AnalyzerReleases.Unshipped.md
@@ -5,15 +5,15 @@
 
 Rule ID | Category | Severity | Notes
 --------|----------|----------|-------
-ALTINN0001 | General | Warning | Project not found
-ALTINN9999 | General | Warning | Unknown error
-ALTINN0010 | Metadata | Warning | App metadata file not found
-ALTINN0011 | Metadata | Warning | App metadata file could not be read
-ALTINN0012 | Metadata | Warning | App metadata file parsing
-ALTINN0013 | Metadata | Warning | App metadata classref not resolved
-ALTINN0014 | Metadata | Warning | App metadata onEntry layout ref not resolved
-ALTINN0040 | Metadata | Warning | Layout-sets file not found
-ALTINN0041 | Metadata | Warning | Layout-sets file could not be read
-ALTINN0042 | Metadata | Warning | Layout-sets file parsing
-ALTINN0043 | Metadata | Warning | Layout-setse data type ref not resolved
-ALTINN0500 | CodeSmells | Warning | CodeSmells
+ALTINNAPP0001 | General | Warning | Project not found
+ALTINNAPP9999 | General | Warning | Unknown error
+ALTINNAPP0010 | Metadata | Warning | App metadata file not found
+ALTINNAPP0011 | Metadata | Warning | App metadata file could not be read
+ALTINNAPP0012 | Metadata | Warning | App metadata file parsing
+ALTINNAPP0013 | Metadata | Warning | App metadata classref not resolved
+ALTINNAPP0014 | Metadata | Warning | App metadata onEntry layout ref not resolved
+ALTINNAPP0040 | Metadata | Warning | Layout-sets file not found
+ALTINNAPP0041 | Metadata | Warning | Layout-sets file could not be read
+ALTINNAPP0042 | Metadata | Warning | Layout-sets file parsing
+ALTINNAPP0043 | Metadata | Warning | Layout-setse data type ref not resolved
+ALTINNAPP0500 | CodeSmells | Warning | CodeSmells
