+++ test/Altinn.App.Analyzers.Tests/DiagnosticsConfigTests.cs
@@ -1,23 +1,13 @@
-// namespace Altinn.App.Analyzers.Tests;
+namespace Altinn.App.Analyzers.Tests;
 
-// public class DiagnosticsConfigTests
-// {
-//     [Fact]
-//     public void All_Diagnostics_Exist()
-//     {
-//         var diagnostics = Diagnostics.All;
+public class DiagnosticsConfigTests
+{
+    [Fact]
+    public void All_Diagnostics_Exist()
+    {
+        var diagnostics = Diagnostics.All;
 
-//         Assert.NotEmpty(Diagnostics.All);
-//         Assert.All(diagnostics, d => d.Id.StartsWith("ALTINN", StringComparison.Ordinal));
-//     }
-
-//     [Fact]
-//     public void MetadataAnalyzer_HasDiagnostics()
-//     {
-//         var analyzer = new MetadataAnalyzer();
-//         var diagnostics = analyzer.SupportedDiagnostics;
-
-//         Assert.NotEmpty(diagnostics);
-//         Assert.All(diagnostics, d => d.Id.StartsWith("ALTINN", StringComparison.Ordinal));
-//     }
-// }
+        Assert.NotEmpty(Diagnostics.All);
+        Assert.All(diagnostics, d => d.Id.StartsWith("ALTINNAPP", StringComparison.Ordinal));
+    }
+}
