+++ test/Altinn.App.Analyzers.Tests/Fixtures/AltinnTestAppFixture.cs
@@ -11,15 +11,6 @@ using Xunit.Abstractions;
 
 namespace Altinn.App.Analyzers.Tests.Fixtures;
 
-// This fixture is used to provide a test app Roslyn workspace for the analyzers to run on.
-// The test app is a real blank Altinn app in the "testapp/" folder.
-// Initializing the fixture is expensive, and can take anywhere between 5-20 seconds on my machine currently,
-// so currently tests run in a "global collection" to avoid re-initializing the fixture for each test.
-// It also gives us some flexibility in that we can make physical changes to project files.
-
-[CollectionDefinition(nameof(AltinnTestAppCollection), DisableParallelization = false)]
-public class AltinnTestAppCollection : ICollectionFixture<AltinnTestAppFixture> { }
-
 public sealed partial class AltinnTestAppFixture : IDisposable
 {
     private ITestOutputHelper? _output;
@@ -53,7 +44,7 @@ public sealed partial class AltinnTestAppFixture : IDisposable
             Assert.True(Directory.Exists(_projectDir));
             var manager = new AnalyzerManager();
             var analyzer = manager.GetProject(Path.Combine(_projectDir, "App.csproj"));
-            _workspace = analyzer.GetWorkspace(addProjectReferences: true);
+            _workspace = analyzer.GetWorkspace();
             Assert.True(_workspace.CanApplyChange(ApplyChangesKind.AddDocument));
             Assert.True(_workspace.CanApplyChange(ApplyChangesKind.RemoveDocument));
             Assert.True(_workspace.CanApplyChange(ApplyChangesKind.ChangeDocument));
