+++ test/Altinn.App.Analyzers.Tests/ModuleInitializer.cs
@@ -1,4 +1,5 @@
 using System.Runtime.CompilerServices;
+using Altinn.App.Analyzers.Tests.Fixtures;
 
 namespace Altinn.App.Analyzers.Tests;
